+++ src/components/pages/apps/network/network-settings-store.ts
@@ -1,0 +1,71 @@
+import { IMarginProps } from '@/components/plot/svg-props'
+import { config } from '@/config'
+import { IDim } from '@/interfaces/dim'
+import { useCallback } from 'react'
+
+import { create } from 'zustand'
+import { createJSONStorage, persist } from 'zustand/middleware'
+
+const SETTINGS_KEY = `${config.appId}:app:network:v2`
+
+const MARGIN = { top: 20, right: 200, bottom: 10, left: 10 }
+
+const PLOT_MARGIN = { top: 20, right: 10, bottom: 100, left: 400 }
+
+export interface INetworkSettings {
+  chargeStrength: number
+  linkDistance: number
+  plot: { size: IDim; margin: IMarginProps }
+}
+
+const DEFAULT_SETTINGS: INetworkSettings = {
+  chargeStrength: -30,
+  linkDistance: 100,
+  plot: {
+    size: { w: 800, h: 600 },
+    margin: { ...PLOT_MARGIN },
+  },
+}
+
+export interface INetworkSettingsStore extends INetworkSettings {
+  updateSettings: (settings: INetworkSettings) => void
+}
+
+export const useNetworkSettingsStore = create<INetworkSettingsStore>()(
+  persist(
+    (set) => ({
+      ...DEFAULT_SETTINGS,
+
+      updateSettings: (settings: INetworkSettings) => {
+        set({
+          ...settings,
+        })
+      },
+    }),
+    {
+      name: SETTINGS_KEY, // name in localStorage
+      storage: createJSONStorage(() => localStorage),
+    }
+  )
+)
+
+export function useNetworkSettings(): {
+  settings: INetworkSettingsStore
+  updateSettings: (settings: INetworkSettingsStore) => void
+  resetSettings: () => void
+} {
+  const settings = useNetworkSettingsStore((state) => state)
+  const updateSettings = useNetworkSettingsStore(
+    (state) => state.updateSettings
+  )
+
+  const resetSettings = useCallback(() => {
+    updateSettings({ ...DEFAULT_SETTINGS })
+  }, [updateSettings])
+
+  return {
+    settings,
+    updateSettings,
+    resetSettings,
+  }
+}