+++ src/components/sortable-item.tsx
@@ -85,8 +85,6 @@ export function BaseSortableItem<T extends ElementType = 'li'>({
   //   transition,
   // }
 
-  console.log(props.tooltip)
-
   const Component = as ?? 'li'
 
   return (